requirements.txt
docling
pymupdf
Pillow
python-doctr
openai
ipywidgets
Fastapi
PyDantic
dependency-injector
paddleocr[all]
paddlepaddle